refactor: 댓글 Dto 클래스 네이밍 변경
- VisitLogDto -> CommentDto
