+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/dto/comment/CommentsResponse.kt
@@ -1,10 +1,9 @@
 package com.woowacourse.staccato.data.dto.comment
 
-import com.woowacourse.staccato.data.dto.moment.VisitLogDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class CommentsResponse(
-    @SerialName("comments") val comments: List<VisitLogDto>
+    @SerialName("comments") val comments: List<CommentDto>
 )
